improve pexpect's output buffering algorithm
current, pexpect's read/search loop chunks of stdout/stderr - 2000
characters at a time by default - and builds a new string buffer for
searching after each `os.read()`;  this results in the allocation of
increasingly large strings.  For a process that outputs 50MB of data,
such as `cat 50.megs.txt`, this allocates strings:

2k, 4k, 6k, ... 50MB

this commit simplifies the string buffering to use a cStringIO.StringIO()
object, drastically improving performance for processes that write large
amounts of data to stdout and stderr.

see: #438